Правки в день перед выступлением (убрали право и испрвавили доступ к страницам)

+++ app/Models/Form.php
@@ -23,7 +23,7 @@ class Form extends Model
     protected $casts = [
         'is_active' => 'boolean',
         'single_entry' => 'boolean',
-        'points' => 'integer'
+        'points' => 'float'
     ];
 
     public function category(): BelongsTo
